requirements.txt
ddgs==9.14.0
langchain_community==0.4.1
langchain_core==1.3.0
langchain_groq==1.1.2
langchain_huggingface==1.2.2
langchain_text_splitters==1.1.2
python-dotenv==1.2.2
